Add Draft-06 and Draft-07 schemas and fix ID scanning
Involved in this fix is CoreSchema is-validated checking and ID scans not
registering $id and $anchor in "properties" as valid IDs. Adding these schemas
exposed a bug where CoreSchema didn't properly check if a schema was already
validated; it now strips the fragment first before checking. It also exposed
that $id and $anchor should not be processed as IDs when inside a "properties",
during an ID scan. Also updating CoreSchema to catch any
MalformedSchemaExceptions thrown while scanning IDs because this assisted in
tracing for finding the problem.