Fix $ref handling when the referenced object contains an $id
The first child $id of the found reference will set the base URI, so the
reference shouldn't do it. Both cases, pointer and plain name, still ensure
the base URI is set. The pointer case does it through the newly updated
followPointer method, which was updated to set the base URI as $id elements
are crossed. As well, it was moved from Validator to ValidatorContext.