Add the concept of "local annotations"
These are used when annotations are not being collected, but they're needed
across sibling keywords. Specifically, "contains" and the "properties"-related
keywords use them. AdditionalItems was changed to not use any annotations and
instead look at the contents of any sibling "items" keywords. UnevaluatedItems
and UnevaluatedProperties now fail with a schema error if annotations are not
being collected. Last, ValidatorContext.getAnnotations was renamed
to annotations.