Improve specification detection
Split the schema determination called from Validator.validate into two methods,
specificationFromSchema and guessSpecification. This gets rid of the boolean
parameter, 'doGuessing'. This is also the first time a keyword is used
directly to remove the need for duplicate logic: CoreSchema handles validating
the virtual $schema directive, i.e. when it's absent and it had to be guessed
or assigned the default.