Fix ValidatorContext.findAndSetRoot to properly handle relative paths
It so happens that Java's URI resolution may not prepend a slash to the path
when resolving a relative path against an absolute URI having an empty path.
This means a URI with a relative path will result in an infinite loop in this
method. This fixes that case and also fixes URL creation so that the path is
properly resolved; the correct constructor is now called.