Add a Validator constructor
This enables storing information across schema validation. This also
necessitated a new way to specify whether annotations and errors are collected:
upon validation, an optional map is passed as a parameter for collecting each
of the respective values. This means that Option.COLLECT_ANNOTATIONS and
Option.COLLECT_ERRORS are now gone.